backend/requirements.txt
fastapi[all]
langchain-experimental
langchain-openai 
pandas 
python-dotenv
tabulate
langchain_google_genai
matplotlib
seaborn
numpy
scikit-learn
scipy